refactor server.ts to multiple files for readablity and to be scalable
