+++ task5/src/services/user-service.ts
@@ -1,0 +1,12 @@
+import { v4 as uuidv4 } from 'uuid';
+import { User } from '../types/user';
+import { users } from '../data/users';
+
+type UserData = Omit<User, 'id'>;
+const USERS = [...users];
+
+export const createUser = (user: UserData) => {
+  const newUser: User = { id: uuidv4(), ...user };
+  USERS.push(newUser);
+  return newUser;
+};
